[mapplauncherd] Avoid signals meant for the control process killing applications. JB#53997
mapplauncherd-qt is being signalled to switch from boot mode to normal
mode with this cmdline:

ExecStart=-/usr/bin/killall -SIGUSR1 booster-qt5

The name booster-qt5 is shared by the boosted applictions which are
forked off that and they are receiving the signal to and because they
don't have a signal handler for it they are being killed. Add a dummy
handler to ignore SIGUSR1 and SIGUSR2 to avoid that happening.